реализовал веб сервер для игры

+++ backend/src/main/java/ru/haliksar/tictactoe/backend/dto/RoomIdDto.java
@@ -6,5 +6,5 @@ import lombok.Data;
 @Data
 @AllArgsConstructor
 public class RoomIdDto {
-    private long roomId;
+    private int roomId;
 }
